requirements.txt
fastapi[all]
asyncpg
sqlalchemy
alembic
python-decouple
celery
redis
passlib
PyJWT
sqladmin
pydantic-settings
